user-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-bigquery
google-auth